requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
snowflake-connector-python>=3.10.0
httpx>=0.27.0
